pyproject.toml
[project]
name = "stackstate-etl"
version = "0.2.0"
description = "StackState Extract-Transform-Load Framework for 4T data ingestion"
authors = [
    {name = "Ravan Naidoo", email = "rnaidoo@stackstate.com"},
]

requires-python = "==3.9"
license = {text = "MIT"}

#######################################################################################################################
# Prod Dependencies
#######################################################################################################################

dependencies = [
    "asteval>=0.9.27",
    "pyyaml>=6.0",
    "schematics>=2.1.1",
    "requests>=2.28.1",
    "six>=1.16.0",
    "importlib-resources>=5.9.0",
    "attrs>=22.1.0",
    "cachetools>=5.2.0",
    "click>=8.1.3",
    "pytz>=2022.2.1",
    "jsonpath-ng>=1.5.3",
]

[project.scripts]
stsetl = "stackstate_etl.cli.main:main"

[build-system]
requires = ["pdm-pep517>=1.0.0"]
build-backend = "pdm.pep517.api"

#######################################################################################################################
# Optional Prod Dependencies
#######################################################################################################################

[project.optional-dependencies]
additional = [
    "pydash>=5.1.1",
    "pandas>=1.5.0",
    "pendulum>=2.1.2",
    "networkx>=2.8.6",
]

[tool]
[tool.pdm]

#######################################################################################################################
# Dev Dependencies
#######################################################################################################################
[tool.pdm.dev-dependencies]
dev = [
    "py-backwards>=0.7",
    "pytest>=7.1.3",
    "pytest-sugar>=0.9.5",
    "toml>=0.10.2",
    "setuptools>=65.4.0",
]
format = [
    "flakehell>=0.9.0",
    "flake8==3.9.0",
    "black>=22.8.0",
    "mypy>=0.971",
    "isort>=5.10.1",
]
#######################################################################################################################
# Scripts
#######################################################################################################################

[tool.pdm.scripts]
pre_build = {composite = ["format", "test"]}
post_build = {composite = ["package_py27"]}
format = {composite = ["isort", "black", "flakehell", "mypy"]}
test = "pytest -s -p no:logging ./tests"
clean = "rm -rf build dist"

#######################################################################################################################
# Helper Scripts
#######################################################################################################################
_.env_file = ".sts.env"
black = "black src tests"
isort = "isort src"
flakehell = "flakehell lint ./src/ ./tests/"
mypy = "mypy src"
package_py27 = {call = "tasks.build:main"}


#######################################################################################################################
# iSort
#######################################################################################################################

[tool.isort]
profile = "black"
multi_line_output = 3
src_paths=["src", "tests"]

#######################################################################################################################
# Black
#######################################################################################################################

[tool.black]
line-length = 120
target-version = ['py37']
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

#######################################################################################################################
# FlakeHell
#######################################################################################################################

[tool.flakehell]
exclude = ["README.rst", "README.md", ".eggs", ".tox", "build",".venv", ".agent"]
include = ["src", "tests"]
format = "colored"
max_line_length = 120
show_source = true
# Temp fix until issue is fixed https://github.com/flakehell/flakehell/issues/10
extended_default_ignore=[]

[tool.flakehell.plugins]
"*" = [
    "+*",
    "-E203",
]

#######################################################################################################################
# PyTest
#######################################################################################################################
[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q"
filterwarnings = [
    "error",
    "ignore::schematics.deprecated.SchematicsDeprecationWarning"
]
pythonpath = [
    "src"
]
testpaths = [
    "tests",
    "integration",
]


#######################################################################################################################
# MyPy
#######################################################################################################################

[tool.mypy]
python_version = 3.7
exclude = [
    '.*test_explore_apis\.py'
]

[[tool.mypy.overrides]]
module = "schematics.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "stackstate_checks.base.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "requests.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "six.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "yaml.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "pytz.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "cachetools.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "asteval.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "jsonpath_ng.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "pandas.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "networkx.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "pendulum.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "pydash.*"
ignore_missing_imports = true
